requirements.txt
Django>=5.2.0
django-filter>=25.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
djangorestframework-simplejwt[crypto]>=5.2.0
drf-yasg>=1.21.0
django-cors-headers>=4.0.0
PyJWT>=2.9.0
psycopg2-binary>=2.9.0
uwsgi>=2.0.0
python-decouple>=3.8
